GasOps_secondusecase_prod_backendcode/requirements.txt
langchain
langchain-community
langchain-openai
azure-identity
azure-search-documents
python-dotenv
fastapi
uvicorn
pdfplumber
pyodbc
pdf2image
pytesseract
azure-ai-documentintelligence
